@@ -23,18 +23,18 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^4.4.3",
-    "@angular/cdk": "^2.0.0-beta.11",
-    "@angular/common": "^4.4.3",
-    "@angular/compiler": "^4.4.3",
-    "@angular/core": "^4.4.3",
-    "@angular/flex-layout": "^2.0.0-beta.9",
-    "@angular/forms": "^4.4.3",
-    "@angular/http": "^4.4.3",
-    "@angular/material": "^2.0.0-beta.11",
-    "@angular/platform-browser": "^4.4.3",
-    "@angular/platform-browser-dynamic": "^4.4.3",
-    "@angular/router": "^4.4.3",
+    "@angular/animations": "^4.4.4",
+    "@angular/cdk": "^2.0.0-beta.12",
+    "@angular/common": "^4.4.4",
+    "@angular/compiler": "^4.4.4",
+    "@angular/core": "^4.4.4",
+    "@angular/flex-layout": "github:angular/flex-layout-builds",
+    "@angular/forms": "^4.4.4",
+    "@angular/http": "^4.4.4",
+    "@angular/material": "^2.0.0-beta.12",
+    "@angular/platform-browser": "^4.4.4",
+    "@angular/platform-browser-dynamic": "^4.4.4",
+    "@angular/router": "^4.4.4",
     "@ngrx/effects": "^4.0.5",
     "@ngrx/entity": "github:ngrx/entity-builds",
     "@ngrx/router-store": "^4.0.4",
@@ -43,45 +43,45 @@
     "@nguniversal/express-engine": "^5.0.0-beta.1",
     "@nguniversal/module-map-ngfactory-loader": "^5.0.0-beta.1",
     "angular-calendar": "^0.21.2",
-    "core-js": "^2.4.1",
+    "core-js": "^2.5.1",
     "crypto-js": "^3.1.9-1",
     "date-fns": "^1.28.5",
-    "express": "^4.15.4",
+    "express": "^4.16.1",
     "hammerjs": "^2.0.8",
     "lodash": "^4.17.4",
-    "rxjs": "^5.4.2",
-    "web-animations-js": "^2.2.5",
-    "zone.js": "^0.8.14"
+    "rxjs": "^5.4.3",
+    "web-animations-js": "^2.3.1",
+    "zone.js": "^0.8.18"
   },
   "devDependencies": {
-    "@angular/cli": "^1.4.3",
-    "@angular/compiler-cli": "^4.4.3",
-    "@angular/language-service": "^4.4.3",
-    "@angular/platform-server": "^4.4.3",
+    "@angular/cli": "^1.4.5",
+    "@angular/compiler-cli": "^4.4.4",
+    "@angular/language-service": "^4.4.4",
+    "@angular/platform-server": "^4.4.4",
     "@angularclass/hmr": "^2.1.3",
-    "@types/crypto-js": "^3.1.33",
-    "@types/express": "^4.0.36",
-    "@types/jasmine": "2.5.53",
-    "@types/jasminewd2": "~2.0.2",
-    "@types/lodash": "^4.14.65",
-    "@types/node": "~6.0.60",
-    "codelyzer": "~3.1.1",
+    "@types/crypto-js": "^3.1.36",
+    "@types/express": "^4.0.37",
+    "@types/jasmine": "2.6.0",
+    "@types/jasminewd2": "~2.0.3",
+    "@types/lodash": "^4.14.77",
+    "@types/node": "~8.0.32",
+    "codelyzer": "~3.2.1",
     "concurrently": "^3.5.0",
-    "jasmine-core": "~2.6.2",
+    "jasmine-core": "~2.8.0",
     "jasmine-marbles": "^0.1.0",
-    "jasmine-spec-reporter": "~4.1.0",
-    "karma": "~1.7.0",
-    "karma-chrome-launcher": "~2.1.1",
+    "jasmine-spec-reporter": "~4.2.1",
+    "karma": "~1.7.1",
+    "karma-chrome-launcher": "~2.2.0",
     "karma-cli": "~1.0.1",
-    "karma-coverage-istanbul-reporter": "^1.2.1",
+    "karma-coverage-istanbul-reporter": "^1.3.0",
     "karma-jasmine": "~1.1.0",
     "karma-jasmine-html-reporter": "^0.2.2",
     "ngrx-store-freeze": "^0.2.0",
     "protractor": "~5.1.2",
     "ts-loader": "^2.3.7",
-    "ts-node": "~3.2.0",
-    "tslint": "~5.3.2",
-    "typescript": "~2.4.2",
+    "ts-node": "~3.3.0",
+    "tslint": "~5.7.0",
+    "typescript": "~2.5.3",
     "webpack-node-externals": "^1.6.0"
   },
   "description": "The Task Manager System",
