@@ -1,8 +1,8 @@
-import {Inject, Injectable} from '@angular/core';
-import {HttpHeaders, HttpClient, HttpParams} from '@angular/common/http';
-import {Observable} from 'rxjs/Observable';
+import { Inject, Injectable } from '@angular/core';
+import { HttpHeaders, HttpClient, HttpParams } from '@angular/common/http';
+import { Observable } from 'rxjs/Observable';
 import * as _ from 'lodash';
-import {Project, User} from '../domain';
+import { Project, User } from '../domain';
 
 @Injectable()
 export class ProjectService {
@@ -10,8 +10,8 @@ export class ProjectService {
   private headers = new HttpHeaders()
     .set('Content-Type', 'application/json');
 
-  constructor(@Inject('BASE_CONFIG') private config: {uri: string},
-              private http: HttpClient) {
+  constructor( @Inject('BASE_CONFIG') private config: { uri: string },
+    private http: HttpClient) {
     // this.headers.append('X-LC-Id', config.LCId);
     // this.headers.append('X-LC-Key', config.LCKey);
   }
@@ -20,7 +20,7 @@ export class ProjectService {
   add(project: Project): Observable<Project> {
     const uri = `${this.config.uri}/${this.domain}`;
     return this.http
-      .post<Project>(uri, JSON.stringify(project), {headers: this.headers});
+      .post<Project>(uri, JSON.stringify(project), { headers: this.headers });
   }
 
   // PUT /projects
@@ -32,7 +32,7 @@ export class ProjectService {
       desc: project.desc
     };
     return this.http
-      .patch<Project>(uri, JSON.stringify(toUpdate), {headers: this.headers});
+      .patch<Project>(uri, JSON.stringify(toUpdate), { headers: this.headers });
   }
 
   // DELETE /projects instead of deleting the records
@@ -40,7 +40,7 @@ export class ProjectService {
     const deltask$ = Observable.from(project.taskLists ? project.taskLists : [])
       .mergeMap(listId => this.http
         .delete(`${this.config.uri}/taskLists/${listId}`))
-        .count();
+      .count();
     const uri = `${this.config.uri}/${this.domain}/${project.id}`;
     return deltask$.switchMap(p => this.http
       .delete(uri)
@@ -53,7 +53,7 @@ export class ProjectService {
     const params = new HttpParams()
       .set('members_like', userId);
     return this.http
-      .get<Project[]>(uri, {params: params, headers: this.headers});
+      .get<Project[]>(uri, { params: params, headers: this.headers });
   }
 
   updateTaskLists(project: Project): Observable<Project> {
@@ -62,7 +62,16 @@ export class ProjectService {
       taskLists: project.taskLists
     };
     return this.http
-      .patch<Project>(uri, JSON.stringify(toUpdate), {headers: this.headers});
+      .patch<Project>(uri, JSON.stringify(toUpdate), { headers: this.headers });
+  }
+
+  insertTaskFilter(project: Project): Observable<Project> {
+    const uri = `${this.config.uri}/${this.domain}/${project.id}`;
+    const toUpdate = {
+      taskFilterId: project.taskFilterId
+    };
+    return this.http
+      .patch<Project>(uri, JSON.stringify(toUpdate), { headers: this.headers });
   }
 
   inviteMembers(projectId: string, users: User[]) {
@@ -74,7 +83,7 @@ export class ProjectService {
         const existingMemberIds = project.members;
         const invitedIds = users.map(user => user.id);
         const newIds = _.union(existingMemberIds, invitedIds);
-        return this.http.patch(uri, JSON.stringify({ members: newIds }), {headers: this.headers});
+        return this.http.patch(uri, JSON.stringify({ members: newIds }), { headers: this.headers });
       });
   }
 }
