@@ -10,8 +10,8 @@ export class TaskmgrPage {
   }
 
   fillInfo() {
-    element(by.id('md-input-0')).sendKeys('dev');
-    element(by.id('md-input-1')).sendKeys('dev');
+    element(by.id('mat-input-0')).sendKeys('dev');
+    element(by.id('mat-input-1')).sendKeys('dev');
     element(by.buttonText('登录')).click();
     return browser.takeScreenshot();
   }
