@@ -1,6 +1,6 @@
-import {Project} from '../domain';
-import {createSelector} from '@ngrx/store';
-import {EntityState, EntityAdapter, createEntityAdapter} from '@ngrx/entity';
+import { Project } from '../domain';
+import { createSelector } from '@ngrx/store';
+import { EntityState, EntityAdapter, createEntityAdapter } from '@ngrx/entity';
 import * as actions from '../actions/project.action';
 
 export interface State extends EntityState<Project> {
@@ -21,18 +21,19 @@ export const initialState: State = adapter.getInitialState({
   selectedId: null
 });
 
-export function reducer (state = initialState, action: actions.Actions): State {
+export function reducer(state = initialState, action: actions.Actions): State {
   switch (action.type) {
     case actions.ADD_SUCCESS:
-      return {...adapter.addOne(action.payload, state), selectedId: null};
+      return { ...adapter.addOne(action.payload, state), selectedId: null };
     case actions.DELETE_SUCCESS:
-      return {...adapter.removeOne(<string>action.payload.id, state), selectedId: null};
+      return { ...adapter.removeOne(<string>action.payload.id, state), selectedId: null };
     case actions.INVITE_SUCCESS:
     case actions.UPDATE_LISTS_SUCCESS:
     case actions.UPDATE_SUCCESS:
-      return {...adapter.updateOne({id: <string>action.payload.id, changes: action.payload}, state), selectedId: null};
+    case actions.INSERT_FILTER_SUCCESS:
+      return { ...adapter.updateOne({ id: <string>action.payload.id, changes: action.payload }, state), selectedId: null };
     case actions.LOADS_SUCCESS:
-      return {...adapter.addMany(action.payload, state), selectedId: null};
+      return { ...adapter.addMany(action.payload, state), selectedId: null };
     case actions.SELECT:
       return { ...state, selectedId: <string>action.payload.id };
     default:
