@@ -4,7 +4,7 @@ import { TaskHistoryVM } from '../../../vm';
 @Component({
   selector: 'app-task-history-item',
   templateUrl: './task-history-item.component.html',
-  styleUrls: ['./task-history-item.component.css'],
+  styleUrls: ['./task-history-item.component.scss'],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
 export class TaskHistoryItemComponent implements OnInit {
